requirements.txt
Django==4.2.4
django-admin-confirm==1.0.0
django-import-export==3.2.0
django-jalali==6.0.1
django-jalali-date==1.0.2
django-nested-admin==4.0.2
httplib2==0.22.0
jdatetime==4.1.1
kavenegar==1.1.2
Pillow==10.0.0
psycopg==3.1.10
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
